refactor: create GenerateTestsTabHelper
