+++ src/main/kotlin/org/jetbrains/research/testspark/display/generatedTestsTab/GeneratedTestsTabData.kt
@@ -1,11 +1,9 @@
-package org.jetbrains.research.testspark.display
+package org.jetbrains.research.testspark.display.generatedTestsTab
 
 import com.intellij.ui.components.JBScrollPane
 import com.intellij.ui.content.Content
 import com.intellij.ui.content.ContentManager
 import org.jetbrains.research.testspark.core.data.TestCase
-import org.jetbrains.research.testspark.display.generatedTestsTab.TestCasePanelFactory
-import org.jetbrains.research.testspark.display.generatedTestsTab.TopButtonsPanelFactory
 import javax.swing.JPanel
 
 class GeneratedTestsTabData {
