+++ src/main/kotlin/org/jetbrains/research/testspark/display/generatedTestsTab/TestCasePanelFactory.kt
@@ -15,7 +15,6 @@ import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.progress.Task
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.ComboBox
-import com.intellij.ui.EditorTextField
 import com.intellij.ui.JBColor
 import com.intellij.ui.LanguageTextField
 import com.intellij.ui.components.JBScrollPane
@@ -29,14 +28,10 @@ import org.jetbrains.research.testspark.core.progress.CustomProgressIndicator
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.data.UIContext
 import org.jetbrains.research.testspark.data.llm.JsonEncoding
-import org.jetbrains.research.testspark.display.GeneratedTestsTabData
 import org.jetbrains.research.testspark.display.coverage.CoverageVisualisationTabFactory
 import org.jetbrains.research.testspark.display.custom.IJProgressIndicator
 import org.jetbrains.research.testspark.display.custom.TestCaseDocumentCreator
-import org.jetbrains.research.testspark.display.utils.IconButtonCreator
-import org.jetbrains.research.testspark.display.utils.ModifiedLinesGetter
-import org.jetbrains.research.testspark.display.utils.ReportUpdater
-import org.jetbrains.research.testspark.display.utils.TestSparkIcons
+import org.jetbrains.research.testspark.display.utils.*
 import org.jetbrains.research.testspark.helpers.JavaClassBuilderHelper
 import org.jetbrains.research.testspark.helpers.LLMHelper
 import org.jetbrains.research.testspark.services.ErrorService
@@ -201,7 +196,7 @@ class TestCasePanelFactory(
             val clipboard: Clipboard = Toolkit.getDefaultToolkit().systemClipboard
             clipboard.setContents(
                 StringSelection(
-                    getEditorTextField(testCase.testName)!!.document.text,
+                    GenerateTestsTabHelper.getEditorTextField(testCase.testName, generatedTestsTabData)!!.document.text,
                 ),
                 null,
             )
@@ -341,7 +336,7 @@ class TestCasePanelFactory(
     private fun addLanguageTextFieldListener(languageTextField: LanguageTextField) {
         languageTextField.document.addDocumentListener(object : DocumentListener {
             override fun documentChanged(event: DocumentEvent) {
-                updateUI()
+                update()
             }
         })
     }
@@ -349,7 +344,7 @@ class TestCasePanelFactory(
     /**
      * Updates the user interface based on the provided code.
      */
-    private fun updateUI() {
+    private fun update() {
         updateTestCaseInformation()
 
         val lastRunCode = lastRunCodes[currentRequestNumber - 1]
@@ -393,8 +388,8 @@ class TestCasePanelFactory(
             testCase.coveredLines = setOf()
         }
 
-        ReportUpdater.updateTestCase(project, report, testCase, coverageVisualisationTabFactory, generatedTestsTabData)
-        update()
+        ReportUpdater.updateTestCase(report, testCase, coverageVisualisationTabFactory, generatedTestsTabData)
+        GenerateTestsTabHelper.update(generatedTestsTabData)
     }
 
     /**
@@ -541,7 +536,7 @@ class TestCasePanelFactory(
         lastRunCodes[currentRequestNumber - 1] = testCase.testCode
 
         SwingUtilities.invokeLater {
-            updateUI()
+            update()
         }
 
         finishProcess()
@@ -565,7 +560,7 @@ class TestCasePanelFactory(
             currentCodes[currentRequestNumber - 1] = testCase.testCode
             lastRunCodes[currentRequestNumber - 1] = testCase.testCode
 
-            updateUI()
+            update()
         }
     }
 
@@ -577,7 +572,7 @@ class TestCasePanelFactory(
             languageTextField.document.setText(lastRunCodes[currentRequestNumber - 1])
             currentCodes[currentRequestNumber - 1] = testCase.testCode
 
-            updateUI()
+            update()
         }
     }
 
@@ -590,60 +585,13 @@ class TestCasePanelFactory(
      * 3. Updating the UI.
      */
     private fun remove() {
-        removeTestCase(testCase.testName)
+        GenerateTestsTabHelper.removeTestCase(testCase.testName, generatedTestsTabData)
 
         runTestButton.isEnabled = false
         isRemoved = true
 
-        ReportUpdater.removeTestCase(project, report, testCase, coverageVisualisationTabFactory, generatedTestsTabData)
-        update()
-    }
-
-    // TODO move to a sep file
-    /**
-     * Updates the user interface of the tool window.
-     *
-     * This method updates the UI of the tool window tab by calling the updateUI
-     * method of the allTestCasePanel object and the updateTopLabels method
-     * of the topButtonsPanel object. It also checks if there are no more tests remaining
-     * and closes the tool window if that is the case.
-     */
-    private fun update() {
-        generatedTestsTabData.allTestCasePanel.updateUI()
-        generatedTestsTabData.topButtonsPanelFactory.update(generatedTestsTabData.testCasePanels, generatedTestsTabData.testsSelected, generatedTestsTabData.testCasePanelFactories)
-
-    }
-
-    // TODO move to a sep file
-    /**
-     * A helper method to remove a test case from the cache and from the UI.
-     *
-     * @param testCaseName the name of the test
-     */
-    private fun removeTestCase(testCaseName: String) {
-        // Update the number of selected test cases if necessary
-        if ((generatedTestsTabData.testCasePanels[testCaseName]!!.getComponent(0) as JCheckBox).isSelected) {
-            generatedTestsTabData.testsSelected--
-        }
-
-        // Remove the test panel from the UI
-        generatedTestsTabData.allTestCasePanel.remove(generatedTestsTabData.testCasePanels[testCaseName])
-
-        // Remove the test panel
-        generatedTestsTabData.testCasePanels.remove(testCaseName)
-    }
-
-    // TODO move to a sep file
-    /**
-     * Retrieve the editor corresponding to a particular test case
-     *
-     * @param testCaseName the name of the test case
-     * @return the editor corresponding to the test case, or null if it does not exist
-     */
-    private fun getEditorTextField(testCaseName: String): EditorTextField? {
-        val middlePanelComponent = generatedTestsTabData.testCasePanels[testCaseName]?.getComponent(2) ?: return null
-        val middlePanel = middlePanelComponent as JPanel
-        return (middlePanel.getComponent(1) as JBScrollPane).viewport.view as EditorTextField
+        ReportUpdater.removeTestCase(report, testCase, coverageVisualisationTabFactory, generatedTestsTabData)
+        GenerateTestsTabHelper.update(generatedTestsTabData)
     }
 
     /**
@@ -701,7 +649,7 @@ class TestCasePanelFactory(
      */
     private fun switchToAnotherCode() {
         languageTextField.document.setText(currentCodes[currentRequestNumber - 1])
-        updateUI()
+        update()
     }
 
     /**
