+++ src/main/kotlin/org/jetbrains/research/testspark/display/utils/ReportUpdater.kt
@@ -1,14 +1,12 @@
 package org.jetbrains.research.testspark.display.utils
 
-import com.intellij.openapi.project.Project
 import org.jetbrains.research.testspark.core.data.Report
 import org.jetbrains.research.testspark.core.data.TestCase
-import org.jetbrains.research.testspark.display.GeneratedTestsTabData
+import org.jetbrains.research.testspark.display.generatedTestsTab.GeneratedTestsTabData
 import org.jetbrains.research.testspark.display.coverage.CoverageVisualisationTabFactory
 
 object ReportUpdater {
     fun updateTestCase(
-        project: Project,
         report: Report,
         testCase: TestCase,
         coverageVisualisationTabFactory: CoverageVisualisationTabFactory,
@@ -21,7 +19,6 @@ object ReportUpdater {
     }
 
     fun removeTestCase(
-        project: Project,
         report: Report,
         testCase: TestCase,
         coverageVisualisationTabFactory: CoverageVisualisationTabFactory,
@@ -33,7 +30,6 @@ object ReportUpdater {
     }
 
     fun unselectTestCase(
-        project: Project,
         report: Report,
         unselectedTestCases: HashMap<Int, TestCase>,
         testCaseId: Int,
@@ -41,11 +37,10 @@ object ReportUpdater {
         generatedTestsTabData: GeneratedTestsTabData,
     ) {
         unselectedTestCases[testCaseId] = report.testCaseList[testCaseId]!!
-        removeTestCase(project, report, report.testCaseList[testCaseId]!!, coverageVisualisationTabFactory, generatedTestsTabData)
+        removeTestCase(report, report.testCaseList[testCaseId]!!, coverageVisualisationTabFactory, generatedTestsTabData)
     }
 
     fun selectTestCase(
-        project: Project,
         report: Report,
         unselectedTestCases: HashMap<Int, TestCase>,
         testCaseId: Int,
