--- src/test/kotlin/org/jetbrains/research/testspark/services/CoverageVisualisationTabFactoryTest.kt
@@ -1,171 +1,0 @@
-package org.jetbrains.research.testspark.services
-
-import com.intellij.openapi.components.service
-import com.intellij.openapi.editor.Editor
-import com.intellij.openapi.wm.RegisterToolWindowTask
-import com.intellij.openapi.wm.ToolWindowAnchor
-import com.intellij.openapi.wm.ToolWindowManager
-import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase
-import org.assertj.core.api.Assertions.assertThat
-import org.evosuite.result.BranchInfo
-import org.evosuite.result.MutationInfo
-import org.evosuite.result.TestGenerationResultImpl
-import org.evosuite.shaded.org.mockito.Mockito
-import org.evosuite.utils.CompactReport
-import org.jetbrains.research.testspark.data.IJReport
-import org.jetbrains.research.testspark.display.coverage.CoverageVisualisationTabFactory
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.TestInstance
-import org.junit.jupiter.params.provider.Arguments
-import java.util.stream.Stream
-
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class CoverageVisualisationTabFactoryTest : LightJavaCodeInsightFixtureTestCase() {
-
-    private lateinit var coverageVisualisationService: CoverageVisualisationTabFactory
-    private val myEditor = Mockito.mock(Editor::class.java)
-
-    private val branch1 = Mockito.mock(BranchInfo::class.java)
-    private val branch2 = Mockito.mock(BranchInfo::class.java)
-    private val branch3 = Mockito.mock(BranchInfo::class.java)
-    private val branch4 = Mockito.mock(BranchInfo::class.java)
-    private val branch5 = Mockito.mock(BranchInfo::class.java)
-
-    private val mutation1 = Mockito.mock(MutationInfo::class.java)
-    private val mutation2 = Mockito.mock(MutationInfo::class.java)
-    private val mutation3 = Mockito.mock(MutationInfo::class.java)
-    private val mutation4 = Mockito.mock(MutationInfo::class.java)
-    private val mutation5 = Mockito.mock(MutationInfo::class.java)
-
-    @BeforeEach
-    override fun setUp() {
-        super.setUp()
-
-        coverageVisualisationService = project.service()
-
-        // Initialise the toolWindow TestSpark
-        ToolWindowManager.getInstance(project)
-            .registerToolWindow(RegisterToolWindowTask("TestSpark", ToolWindowAnchor.RIGHT))
-
-        // Disable coverage visualisation
-//        val settingsProjectState = ApplicationManager.getApplication().getService(SettingsProjectService::class.java).state
-//        settingsProjectState.showCoverageCheckboxSelected = false
-    }
-
-    @AfterEach
-    override fun tearDown() {
-        super.tearDown()
-    }
-
-    // TODO uncomment
-//    @ParameterizedTest
-//    @MethodSource("valueGenerator")
-//    fun fillToolWindowContents(
-//        className: String,
-//        coveredLinesSet: Set<Int>,
-//        uncoveredLinesSet: Set<Int>,
-//        coveredBranchesSet: Set<BranchInfo>,
-//        uncoveredBranchesSet: Set<BranchInfo>,
-//        coveredMutationSet: Set<MutationInfo>,
-//        uncoveredMutationSet: Set<MutationInfo>,
-//        lineCoverage: String,
-//        branchCoverage: String,
-//        mutationCoverage: String,
-//    ) {
-//        val report = Report(CompactReport(TestGenerationResultImpl()))
-//        report.UUT = className
-//        report.allCoveredLines = coveredLinesSet
-//        report.allUncoveredLines = uncoveredLinesSet
-//        report.allCoveredBranches = coveredBranchesSet
-//        report.allUncoveredBranches = uncoveredBranchesSet
-//        report.allCoveredMutation = coveredMutationSet
-//        report.allUncoveredMutation = uncoveredMutationSet
-//
-//        coverageVisualisationService.showCoverage(report, myEditor)
-//        assertThat(className).isEqualTo(coverageToolWindowDisplayService.data[0])
-//        assertThat(lineCoverage).isEqualTo(coverageToolWindowDisplayService.data[1])
-//        assertThat(branchCoverage).isEqualTo(coverageToolWindowDisplayService.data[2])
-//        assertThat(mutationCoverage).isEqualTo(coverageToolWindowDisplayService.data[3])
-//    }
-
-    fun testCreateToolWindowTabTestSingleContent() {
-        coverageVisualisationService.show(IJReport(CompactReport(TestGenerationResultImpl())))
-        val toolWindow = ToolWindowManager.getInstance(project).getToolWindow("TestSpark")!!
-
-        // Verify only 1 content is created
-        assertThat(toolWindow.contentManager.contents.size).isEqualTo(1)
-    }
-
-    fun testCreateToolWindowTabTestContent() {
-        coverageVisualisationService.show(IJReport(CompactReport(TestGenerationResultImpl())))
-        val toolWindow = ToolWindowManager.getInstance(project).getToolWindow("TestSpark")!!
-        val content = toolWindow.contentManager.getContent(0)!!
-        assertThat(content.displayName).isEqualTo("Coverage")
-    }
-
-    private fun valueGenerator(): Stream<Arguments> {
-        return Stream.of(
-            Arguments.of(
-                "MyClass",
-                setOf(1, 2),
-                setOf(2, 3),
-                setOf(branch1),
-                setOf(branch2),
-                setOf(mutation1, mutation2),
-                setOf<MutationInfo>(),
-                "50% (2/4)",
-                "50% (1/2)",
-                "100% (2/2)",
-            ),
-            Arguments.of(
-                "MyClass",
-                setOf(1, 2, 5),
-                setOf(2, 3),
-                setOf<BranchInfo>(),
-                setOf(branch1, branch2),
-                setOf(mutation1, mutation2),
-                setOf(mutation3),
-                "60% (3/5)",
-                "0% (0/2)",
-                "67% (2/3)",
-            ),
-            Arguments.of(
-                "MyClass",
-                setOf(1),
-                setOf(2, 3, 4, 5, 6),
-                setOf(branch5),
-                setOf(branch1, branch2, branch3, branch4),
-                setOf(mutation1),
-                setOf(mutation3),
-                "17% (1/6)",
-                "20% (1/5)",
-                "50% (1/2)",
-            ),
-            Arguments.of(
-                "MyClass",
-                setOf(1, 2, 3, 4, 5, 6, 7),
-                setOf<Int>(),
-                setOf(branch1, branch2, branch3, branch4, branch5),
-                setOf<BranchInfo>(),
-                setOf<MutationInfo>(),
-                setOf(mutation1, mutation3),
-                "100% (7/7)",
-                "100% (5/5)",
-                "0% (0/2)",
-            ),
-            Arguments.of(
-                "MyClass",
-                setOf<Int>(),
-                setOf(1, 2, 3),
-                setOf(branch1, branch2, branch3, branch4, branch5),
-                setOf(branch1, branch2, branch3),
-                setOf(mutation1, mutation2),
-                setOf(mutation1, mutation2, mutation3, mutation4, mutation5),
-                "0% (0/3)",
-                "63% (5/8)",
-                "29% (2/7)",
-            ),
-        )
-    }
-}
